third_party/WebKit/LayoutTests/fast/events/focus-event-source-device-from-mouse-expected.txt

Test that focus events from a mouse device should have a sourceDevice and its firesTouchEvents should be false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


tests on div
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

tests on input
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

tests on button
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

tests on textarea
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

tests on select
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

tests on a
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

PASS successfullyParsed is true

TEST COMPLETE
